vagrant: drop the Red Hat box selection menu
Since we now accept a freeform string for Fedora boxes, rename it to
be a more generic "Fedora-derived distro", and require a boxname. With
this, we can drop the whole RHEL section and treat it as an extension of
the Fedora "family".

If we later get around to implementing registration for RHEL machines,
we can add new Kconfig options as necessary.

Signed-off-by: Jeff Layton <jlayton@kernel.org>